Commit: Forgot to commit hence so many edits. Testing for JobApplication complete. Postman works and all unit tests passed. In Progress for User Model and related classes.

+++ src/main/java/com/dave/jobtracker/job_application_tracker/repository/JARepository.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface JARepository extends JpaRepository<JobApplication, Long>{
+public interface JARepository extends JpaRepository<JobApplication, Integer>{
 
     List<JobApplication> findByDateApplied(LocalDate dateApplied);
